% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_standardizing_script.R
\name{concat_postal_codes}
\alias{concat_postal_codes}
\title{Concatenate Postal Codes}
\usage{
concat_postal_codes(df, alt_postal_codes)
}
\arguments{
\item{df}{The current built up and processed data frame containing the column alt_postal_code.}

\item{alt_postal_codes}{A vector of alternative postal codes that will first be added to the data frame column before de-duplication takes place.}
}
\value{
The original data frame with an additional column called "alt_postal_codes" containing the de-duplicated extracted codes.
}
\description{
Modifies the alt_postal_code column of the data frame, by splitting the data and
removing any duplicate postal codes that it may contain in a single row.
}
\examples{
df <- concat_postal_codes(df, c("A1A1A1", NA, "B1B1B1"))
}
