% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization_script.R
\name{error_handle}
\alias{error_handle}
\title{Error Handle}
\usage{
error_handle(
  standardization_rules_metadata,
  err_msg,
  cleaned_file,
  error_file_path,
  error_folder_path
)
}
\arguments{
\item{standardization_rules_metadata}{A database connection that is used in addition to the dataset_id, passed from the standardize_data() function. Is also used in the pre_process_data(), and pre_process_chunks() functions.}

\item{err_msg}{An error message indicated what went wrong in the program, this message will be written to an output .txt file and placed in a folder containing error messages}

\item{cleaned_file}{A database connection to the sqlite file holding the currently processed dataset that was being written to at the time of error.}

\item{error_file_path}{A path to the input file.}

\item{error_folder_path}{A path to the output folder where the error .txt file will be created.}
}
\description{
The error_handle() function is an exception/error handling function that will log error messages and print
these messages to an output log, allowing the user to read and see what went wrong in the program, allowing
them to hopefully understand and make corrections to a specific datasets metadata
}
\examples{
error_handle(db, "Invalid Dataset Code - Stopping Program", clean_db, "path/to/file.txt", "path/to/folder");
}
