% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_standardizing_script.R
\name{extract_postal_codes}
\alias{extract_postal_codes}
\title{Extract Postal Codes}
\usage{
extract_postal_codes(to_extract, df)
}
\arguments{
\item{to_extract}{A vector of data which may contain concatenated postal codes which should be extracted.}

\item{df}{The current built up and processed data frame, a new column called “alt_postal_code” is defined and the extracted postal codes are appended to it.}
}
\value{
A list of two values, one being the original field under "orig_field", the other being the extracted postal codes under "p_codes"
}
\description{
Takes in a vector of values that may potentially have postal codes embedded in them, regular expressions are used
to extract the values and append them to a new column in the current processed data frame called “alt_postal_code”.
}
\examples{
return_list <- extract_postal_codes(c("123 Fake Address A1A1A1", "No Postal Code", "B1B 1B1 321 Fake Address"), source_df)
}
