% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_standardizing_script.R
\name{list_curr_given_names}
\alias{list_curr_given_names}
\title{List Current Given Names}
\usage{
list_curr_given_names(names_to_append, df)
}
\arguments{
\item{names_to_append}{A vector of given names to be concatenated together, separated by a space. Contains all current processed given names of the dataset.}

\item{df}{The current built up and processed data frame, a new column called “all_curr_given_names” is defined and the processed names are appended to it.}
}
\value{
The original data frame with an additional column called "all_curr_given_names" with names_to_append pasted to it.
}
\description{
list_curr_given_names() will take in a vector of primary of secondary given names, the current built up data frame,
and append all given names if the user flag is defined as yes before pre-processing.
}
\examples{
df <- list_curr_given_names(c("Maria", "Larry", "John"), df)
}
