% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_standardizing_script.R
\name{list_curr_names}
\alias{list_curr_names}
\title{List All Current Names}
\usage{
list_curr_names(names_to_append, df)
}
\arguments{
\item{names_to_append}{A vector of names to be concatenated together, separated by a space. Contains all current processed names of the dataset.}

\item{df}{The current built up and processed data frame, a new column called “all_curr_names” is defined and the processed names are appended to it.}
}
\value{
The original data frame with an additional column called "all_curr_names" with names_to_append pasted to it.
}
\description{
list_curr_names() will take in a vector of names, the current built up data frame, and append all current names (primary/secondary given names,
primary/secondary surnames, and prior surnames) if the user flag is defined as yes before pre-processing.
}
\examples{
df <- list_curr_names(c("James", "Jane", "Cole"), df)
}
