% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_standardizing_script.R
\name{list_curr_surnames}
\alias{list_curr_surnames}
\title{List Current Surnames}
\usage{
list_curr_surnames(names_to_append, df)
}
\arguments{
\item{names_to_append}{A vector of surnames to be concatenated together, separated by a space. Contains all current processed surnames of the dataset.}

\item{df}{The current built up and processed data frame, a new column called “all_curr_surnames” is defined and the processed names are appended to it.}
}
\value{
The original data frame with an additional column called "all_curr_surnames" with names_to_append pasted to it.
}
\description{
list_curr_surnames() will take in a vector of primary of secondary surnames, the current built up data frame,
and append all surnames if the user flag is defined as yes before pre-processing.
}
\examples{
df <- list_curr_surnames(c("Doe", "Doe", "Chuchmach"), df)
}
