% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization_script.R
\name{pre_process_chunks}
\alias{pre_process_chunks}
\title{Pre Process Chunks}
\usage{
pre_process_chunks(
  file_path,
  dataset_metadata_conn,
  dataset_id,
  flag_lookup_table,
  output_folder,
  dataset_code
)
}
\arguments{
\item{file_path}{A path to a file.}

\item{dataset_metadata_conn}{A database connection that is used in addition to the dataset_id, passed from the standardize_data() function. Is also used in the pre_process_data() function.}

\item{dataset_id}{A dataset_id found and passed from the standardize_data() function, is used to make various calls to the database connection argument to obtain information required to pre-process the health data.}

\item{flag_lookup_table}{A lookup table containing manually set or default values used by the flag script which will use the values of the flags to determine how to pre-process and output the health data.}

\item{output_folder}{An output folder that the successfully cleaned data ready for linkage will be placed in once pre-processing is complete.}

\item{dataset_code}{A dataset code used as prefix for the successful output and error files.}
}
\value{
A clean database connection with a single table called "cleaned_data_table" which contains the processed fields from the input file.
}
\description{
The pre_process_chunks() function takes in a path to a file, metadata, the
dataset_id of what dataset we're standardizing, as well as pre_processing flags
used in the pre_process_data function, and an output folder to move the
standardized file to.
}
\examples{
clean_db_conn <- pre_process_chunks("path/to/file.txt", db, 1, flag_lookup, "path/to/folder", "fakecode")
}
