% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_standardizing_script.R
\name{split_compound_field}
\alias{split_compound_field}
\title{Split Compound Field}
\usage{
split_compound_field(compound_field, split_objects, split_type)
}
\arguments{
\item{compound_field}{A vector of compound data meant to be split using the objects in “split_objects”, the other argument “split_type” will determine the algorithm/logic for splitting a compound field.}

\item{split_objects}{A vector of split objects that will either use split logic to separate a compound field based on a specific character or string. Or we will use substr logic to separate a compound field based on integer indexes.}

\item{split_type}{Splitting a compound field will either be of type separators, or of type indexes.}
}
\value{
A list of split values.
}
\description{
split_compound_field() will take in a vector of single fields that are to be split into separate fields,
it will either use separators or indexes based on the source fields metadata.
}
\examples{
split_dates <- split_compound_field(c("1960-01-01", "2024-04-04"), c("-", "-"), "separators")
}
