% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization_script.R
\name{standardize_data}
\alias{standardize_data}
\title{Standardize Data}
\usage{
standardize_data(
  input_file_path,
  input_dataset_code,
  input_flags,
  output_folder,
  standardization_rules_metadata
)
}
\arguments{
\item{input_file_path}{A path to a file.}

\item{input_dataset_code}{A dataset_code, used in the query statement within the function which will get the desired dataset code. File names meant to be pre-processed should contain a prefix containing the dataset code kept in the metadata, separated by something like a hyphen, underscore, etc.}

\item{input_flags}{A lookup table containing manually set or default values used by the flag script which will use the values of the flags to determine how to pre-process and output the health data.}

\item{output_folder}{An output folder that the successfully cleaned data ready for linkage will be placed in once pre-processing is complete.}

\item{standardization_rules_metadata}{A path to a standardization rules metadata file.}
}
\value{
A cleaned data frame containing the cleaned table from pre_process_chunks() returned database (as long as file size is <1GB)
}
\description{
The standardize data function will open a database connection to a SQLite file which contains all metadata info
of the datasets that will be standardized.It will then fetch a dataset_id using a selected dataset_code and pass
it to a chunking script which will read and process the file.
}
\examples{
clean_df <- standardize_data("path/to/file.txt", "fakecode", flag_lookup, "path/to/folder", "path/to/metadata.sqlite")
}
