% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_standardizing_script.R
\name{standardize_locations}
\alias{standardize_locations}
\title{Standardize Addresses}
\usage{
standardize_locations(input_locations, flag_lookup_table)
}
\arguments{
\item{flag_lookup_table}{A lookup table containing manually set or default values that will determine how persons addresses are handled specifically by the function.}

\item{input_addresses}{A vector of addresses, can be primary or alternative addresses, countries, cities, or any location based field.}
}
\value{
A vector of processed address/location based values.
}
\description{
The standardize_addresses() function serves as a general/abstract function that will use input flags
and data standards to standardize fields in the data set relating to a person's address or location.
}
\examples{
processed_locations <- standardize_locations(c("123 Fake St", "canada", "winniPEG"), flag_lookup)
}
