% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_standardizing_script.R
\name{standardize_names}
\alias{standardize_names}
\title{Standardize Names}
\usage{
standardize_names(input_names, flag_lookup_table)
}
\arguments{
\item{input_names}{A vector of names, can be primary or secondary given names, primary or secondary surnames, and prior surnames.}

\item{flag_lookup_table}{A lookup table containing manually set or default values that will determine how persons names are handled specifically by the function.}
}
\value{
A vector of processed names.
}
\description{
The standardize_names() function serves as a general/abstract function that will
use input flags and data standards to standardize fields in the data set relating
to a person's name.
}
\examples{
processed_names <- standardize_names(c("John", "Jane", "Cole"), flag_lookup)
}
