% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization_ui.R
\name{startDataStandardizationUI}
\alias{startDataStandardizationUI}
\title{Start Standardization UI}
\usage{
startDataStandardizationUI()
}
\description{
Will start the local shiny app upon call, from here the user can use the standardizing function
in a user friendly way by way of the GUI.
}
\examples{
startDataStandardizationUI()
}
